(PC-7905): Move queries to core
