+++ src/pcapi/local_providers/install.py
@@ -1,7 +1,7 @@
 from pcapi.core.providers.models import Provider
+from pcapi.core.providers.repository import get_provider_by_local_class
 import pcapi.local_providers
 from pcapi.models.db import db
-from pcapi.repository.provider_queries import get_provider_by_local_class
 
 
 def install_local_providers():
