+++ src/pcapi/scheduled_tasks/titelive_clock.py
@@ -1,10 +1,10 @@
 from apscheduler.schedulers.blocking import BlockingScheduler
 
 from pcapi.core.logging import install_logging
+from pcapi.core.providers.repository import get_provider_by_local_class
 from pcapi.local_providers.provider_manager import synchronize_data_for_provider
 from pcapi.local_providers.venue_provider_worker import update_venues_for_specific_provider
 from pcapi.models.feature import FeatureToggle
-from pcapi.repository.provider_queries import get_provider_by_local_class
 from pcapi.scheduled_tasks import utils
 from pcapi.scheduled_tasks.decorators import cron_context
 from pcapi.scheduled_tasks.decorators import cron_require_feature
