+++ tests/local_providers/titelive_things_test.py
@@ -3,6 +3,7 @@ from unittest.mock import patch
 
 import pytest
 
+from pcapi.core.providers.repository import get_provider_by_local_class
 from pcapi.local_providers import TiteLiveThings
 from pcapi.model_creators.generic_creators import create_booking
 from pcapi.model_creators.generic_creators import create_offerer
@@ -18,7 +19,6 @@ from pcapi.models import Offer
 from pcapi.models import Product
 from pcapi.models.local_provider_event import LocalProviderEventType
 from pcapi.repository import repository
-from pcapi.repository.provider_queries import get_provider_by_local_class
 
 
 class TiteliveThingsTest:
